Fixed output image format bug

+++ src/primitives/color.rs
@@ -25,9 +25,9 @@ pub fn write_color<W: Write>(out: &mut W, pixel_color: &Color) -> io::Result<()>
 
     // Translate the color components from the range [0,1] to [0,255].
     let intensity: Interval = Interval::new(0.000, 0.999);
-    let rbyte = 256.0 * intensity.clamp(r);
-    let gbyte = 256.0 * intensity.clamp(g);
-    let bbyte = 256.0 * intensity.clamp(b);
+    let rbyte = (256.0 * intensity.clamp(r)) as u32;
+    let gbyte = (256.0 * intensity.clamp(g)) as u32;
+    let bbyte = (256.0 * intensity.clamp(b)) as u32;
 
     // Escribir los componentes del color del píxel.
     writeln!(out, "{} {} {}", rbyte, gbyte, bbyte)
